pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Starter template for making a new agent LangGraph."
authors = [
    { name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.11.11",
    "beautifulsoup4>=4.12.3",
    "bs4>=0.0.2",
    "html2text>=2024.2.26",
    "httpx>=0.28.1",
    "langchain>=0.3.14",
    "langchain-anthropic>=0.3.1",
    "langchain-community>=0.3.14",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.6",
    "lxml>=5.3.0",
    "markitdown>=0.0.1a1",
    "openai>=1.59.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scrapy>=2.12.0",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
